refactor: split the user groups page into a rack and its dialogs
pages/teams/Groups.vue goes from 514 to 66 lines as max-lines counts them, and its entry leaves the exemption block in eslint.config.js. The creation form, the searchable rack, the editor with its member picker and the deletion confirmation each move to a component under components/teams/, every one owning the form only it submits. The rendered markup is unchanged but for the two wrappers the extraction needs.
