+++ eslint.config.js
@@ -258,7 +258,6 @@ export default defineConfigWithVueTs(
             'resources/js/layouts/MainLayout.vue',
             'resources/js/pages/teams/Analytics.vue',
             'resources/js/pages/teams/AuditExports.vue',
-            'resources/js/pages/teams/Groups.vue',
         ],
         rules: {
             'max-lines': 'off',
