+++ resources/js/pages/teams/Groups.vue
@@ -1,39 +1,20 @@
 <script setup lang="ts">
-import { Head, useForm } from '@inertiajs/vue3';
-import { Pencil, Plus, Search, Trash2, Users, X } from '@lucide/vue';
-import { computed, ref, watch } from 'vue';
-import FormField from '@/components/FormField.vue';
+import { Head } from '@inertiajs/vue3';
+import { ref } from 'vue';
 import Heading from '@/components/Heading.vue';
-import { Button } from '@/components/ui/button';
-import {
-    Dialog,
-    DialogClose,
-    DialogContent,
-    DialogDescription,
-    DialogFooter,
-    DialogHeader,
-    DialogTitle,
-} from '@/components/ui/dialog';
-import { Input } from '@/components/ui/input';
-import { useInitials } from '@/composables/useInitials';
+import CreateUserGroupForm from '@/components/teams/CreateUserGroupForm.vue';
+import DeleteUserGroupDialog from '@/components/teams/DeleteUserGroupDialog.vue';
+import EditUserGroupDialog from '@/components/teams/EditUserGroupDialog.vue';
+import UserGroupsRack from '@/components/teams/UserGroupsRack.vue';
 import { translate } from '@/lib/i18n';
 import { edit, index } from '@/routes/teams';
-import {
-    destroy,
-    index as groupsIndex,
-    store,
-    update,
-} from '@/routes/teams/groups';
-import {
-    destroy as removeMember,
-    store as addMember,
-} from '@/routes/teams/groups/members';
+import { index as groupsIndex } from '@/routes/teams/groups';
 import type { Team } from '@/types';
 
 type UserGroup = App.Data.UserGroupData;
 type Member = App.Data.UserData;
 
-const props = defineProps<{
+defineProps<{
     team: Team;
     groups: UserGroup[];
     members: Member[];
@@ -53,154 +34,11 @@ defineOptions({
     }),
 });
 
-const { getInitials } = useInitials();
+/** The editor, opened through its exposed handler so it can seed its form. */
+const editor = ref<InstanceType<typeof EditUserGroupDialog> | null>(null);
 
-/** The row's member-count label, e.g. "3 members". */
-function memberCountLabel(group: UserGroup): string {
-    return group.membersCount === 1
-        ? translate(':count member', { count: group.membersCount })
-        : translate(':count members', { count: group.membersCount });
-}
-
-const createForm = useForm({ name: '', slug: '' });
-
-function submitCreate(): void {
-    createForm.post(store(props.team.slug).url, {
-        preserveScroll: true,
-        onSuccess: () => createForm.reset(),
-    });
-}
-
-const search = ref('');
-const filteredGroups = computed<UserGroup[]>(() => {
-    const needle = search.value.trim().replace(/^@/, '').toLowerCase();
-
-    if (needle === '') {
-        return props.groups;
-    }
-
-    return props.groups.filter(
-        (group) =>
-            group.slug.includes(needle) ||
-            group.name.toLowerCase().includes(needle),
-    );
-});
-
-const editing = ref<UserGroup | null>(null);
-const editForm = useForm({ name: '', slug: '' });
-const memberSearch = ref('');
-
-function openEditor(group: UserGroup): void {
-    editing.value = group;
-    editForm.defaults({ name: group.name, slug: group.slug });
-    editForm.reset();
-    editForm.clearErrors();
-    memberSearch.value = '';
-}
-
-/**
- * The editor renders off the `groups` prop, so after every membership change it
- * has to re-read the freshly reloaded group rather than the stale snapshot it
- * was opened with.
- */
-watch(
-    () => props.groups,
-    (groups) => {
-        if (editing.value === null) {
-            return;
-        }
-
-        editing.value =
-            groups.find((group) => group.id === editing.value?.id) ?? null;
-    },
-);
-
-function submitRename(): void {
-    const group = editing.value;
-
-    if (group === null) {
-        return;
-    }
-
-    editForm.patch(update({ team: props.team.slug, group: group.id }).url, {
-        preserveScroll: true,
-    });
-}
-
-const membershipForm = useForm<{ user_id: string }>({ user_id: '' });
-
-const candidates = computed<Member[]>(() => {
-    const group = editing.value;
-
-    if (group === null) {
-        return [];
-    }
-
-    const already = new Set(group.members.map((member) => member.id));
-    const needle = memberSearch.value.trim().toLowerCase();
-
-    return props.members
-        .filter(
-            (member) =>
-                !already.has(member.id) &&
-                (needle === '' || member.name.toLowerCase().includes(needle)),
-        )
-        .slice(0, 8);
-});
-
-function addToGroup(member: Member): void {
-    const group = editing.value;
-
-    if (group === null) {
-        return;
-    }
-
-    membershipForm.user_id = member.id;
-    membershipForm.post(
-        addMember({ team: props.team.slug, group: group.id }).url,
-        { preserveScroll: true, onSuccess: () => (memberSearch.value = '') },
-    );
-}
-
-function removeFromGroup(memberId: string): void {
-    const group = editing.value;
-
-    if (group === null) {
-        return;
-    }
-
-    membershipForm.delete(
-        removeMember({
-            team: props.team.slug,
-            group: group.id,
-            user: memberId,
-        }).url,
-        { preserveScroll: true },
-    );
-}
-
+/** The group the deletion dialog is confirming, or `null` while it is closed. */
 const pendingRemoval = ref<UserGroup | null>(null);
-const removalForm = useForm({});
-
-function confirmRemoval(): void {
-    const group = pendingRemoval.value;
-
-    if (group === null) {
-        return;
-    }
-
-    removalForm.delete(
-        destroy({ team: props.team.slug, group: group.id }).url,
-        {
-            preserveScroll: true,
-            // Closed only once the delete lands: a failed request leaves the
-            // dialog open rather than dismissing it as though it had worked.
-            onSuccess: () => {
-                pendingRemoval.value = null;
-            },
-        },
-    );
-}
 </script>
 
 <template>
@@ -217,351 +55,25 @@ function confirmRemoval(): void {
             "
         />
 
-        <form
+        <CreateUserGroupForm
             v-if="permissions.canManageUserGroups"
-            data-test="group-create-form"
-            class="flex flex-col gap-3 rounded-xl border border-dashed border-border p-4 sm:flex-row sm:items-start"
-            @submit.prevent="submitCreate"
-        >
-            <div
-                class="flex size-10 shrink-0 items-center justify-center rounded-xl border border-border bg-muted text-muted-foreground"
-            >
-                <Users class="size-4" />
-            </div>
-            <div class="flex flex-1 flex-col gap-3">
-                <div class="flex flex-col gap-1">
-                    <span class="text-sm font-semibold">{{
-                        $t('Create a group')
-                    }}</span>
-                    <span class="text-xs text-muted-foreground">{{
-                        $t(
-                            'The handle is what people type after @. Leave it blank to derive it from the name.',
-                        )
-                    }}</span>
-                </div>
-                <!-- The row heading above names the pair, so each field's own
-                     label is hidden rather than repeated. It stays a real
-                     `<label for>` all the same: a placeholder disappears the
-                     moment someone types, which leaves the field unnamed
-                     exactly when they might look for the name again.
+            :team="team.slug"
+        />
 
-                     The error's space is not reserved here. These cells are
-                     narrow enough that a message wraps to three lines, and the
-                     row is the last thing in the card — so drawn out of flow it
-                     would run past the card's own border, and there is nothing
-                     below it to be pushed around by letting it grow instead. -->
-                <div class="flex flex-col gap-3 sm:flex-row sm:items-start">
-                    <FormField
-                        :label="$t('Group name')"
-                        label-class="sr-only"
-                        :error="createForm.errors.name"
-                        :reserve="false"
-                        v-slot="{ id }"
-                    >
-                        <Input
-                            :id="id"
-                            v-model="createForm.name"
-                            data-test="group-name-input"
-                            :placeholder="$t('Dev Team')"
-                            class="max-md:h-11 sm:w-56"
-                            autocomplete="off"
-                        />
-                    </FormField>
-                    <FormField
-                        :label="$t('Group handle')"
-                        label-class="sr-only"
-                        :error="createForm.errors.slug"
-                        :reserve="false"
-                        v-slot="{ id }"
-                    >
-                        <Input
-                            :id="id"
-                            v-model="createForm.slug"
-                            data-test="group-slug-input"
-                            placeholder="@dev-team"
-                            class="font-mono max-md:h-11 sm:w-48"
-                            autocapitalize="off"
-                            autocomplete="off"
-                            spellcheck="false"
-                        />
-                    </FormField>
-                    <Button
-                        type="submit"
-                        data-test="group-create-button"
-                        class="rounded-full max-md:h-11"
-                        :disabled="createForm.processing"
-                    >
-                        <Plus class="size-4" /> {{ $t('Create') }}
-                    </Button>
-                </div>
-            </div>
-        </form>
-
-        <div class="relative">
-            <Search
-                class="pointer-events-none absolute top-1/2 left-3 size-4 -translate-y-1/2 text-muted-foreground"
-            />
-            <Input
-                v-model="search"
-                data-test="group-search"
-                :placeholder="$t('Search groups')"
-                class="rounded-full pl-9 max-md:h-11"
-            />
-        </div>
-
-        <p
-            v-if="filteredGroups.length === 0"
-            data-test="group-empty"
-            class="text-sm text-muted-foreground"
-        >
-            {{ $t('No user groups yet.') }}
-        </p>
-
-        <ul v-else class="space-y-2" data-test="group-list">
-            <li
-                v-for="group in filteredGroups"
-                :key="group.id"
-                class="flex items-center gap-3 rounded-xl border border-border bg-card p-3 max-md:flex-wrap"
-                :data-test="`group-row-${group.slug}`"
-            >
-                <div
-                    class="flex size-8 shrink-0 items-center justify-center rounded-lg border border-border bg-muted text-muted-foreground"
-                >
-                    <Users class="size-4" aria-hidden="true" />
-                </div>
-                <div class="flex min-w-0 flex-1 flex-col max-md:basis-3/5">
-                    <span
-                        class="truncate font-mono text-sm font-semibold text-foreground"
-                        >@{{ group.slug }}</span
-                    >
-                    <span class="truncate text-xs text-muted-foreground"
-                        >{{ group.name
-                        }}<span class="md:hidden">
-                            · {{ memberCountLabel(group) }}</span
-                        ></span
-                    >
-                </div>
-                <span
-                    class="w-28 shrink-0 text-xs text-muted-foreground max-md:hidden"
-                    >{{ memberCountLabel(group) }}</span
-                >
-                <!-- Icon actions rather than inline text links: the destructive
-                     token at 12px does not clear 4.5:1 on `bg-card` in the dark
-                     theme, while an icon only owes the 3:1 graphics threshold. -->
-                <div
-                    v-if="permissions.canManageUserGroups"
-                    class="flex shrink-0 items-center gap-1 max-md:ml-auto"
-                >
-                    <Button
-                        variant="ghost"
-                        size="icon-sm"
-                        type="button"
-                        :data-test="`group-edit-${group.slug}`"
-                        :aria-label="$t('Edit :name', { name: group.name })"
-                        class="shrink-0 rounded-full max-md:size-11"
-                        @click="openEditor(group)"
-                    >
-                        <Pencil class="size-4" />
-                    </Button>
-                    <Button
-                        variant="ghost"
-                        size="icon-sm"
-                        type="button"
-                        :data-test="`group-remove-${group.slug}`"
-                        :aria-label="$t('Delete :name', { name: group.name })"
-                        class="shrink-0 rounded-full text-destructive-text max-md:size-11"
-                        @click="pendingRemoval = group"
-                    >
-                        <Trash2 class="size-4" />
-                    </Button>
-                </div>
-            </li>
-        </ul>
+        <UserGroupsRack
+            :groups="groups"
+            :can-manage="permissions.canManageUserGroups"
+            @edit="editor?.open($event)"
+            @remove="pendingRemoval = $event"
+        />
     </div>
 
-    <Dialog
-        :open="editing !== null"
-        @update:open="(open) => !open && (editing = null)"
-    >
-        <DialogContent data-test="group-edit-dialog">
-            <DialogHeader>
-                <DialogTitle>{{ $t('Edit group') }}</DialogTitle>
-                <DialogDescription>
-                    {{
-                        $t(
-                            'Renaming a group never rewrites messages already sent. They keep the handle they were written with.',
-                        )
-                    }}
-                </DialogDescription>
-            </DialogHeader>
+    <EditUserGroupDialog
+        ref="editor"
+        :team="team.slug"
+        :groups="groups"
+        :members="members"
+    />
 
-            <form
-                class="flex flex-col gap-3 sm:flex-row sm:items-start"
-                data-test="group-rename-form"
-                @submit.prevent="submitRename"
-            >
-                <FormField
-                    :label="$t('Group name')"
-                    label-class="sr-only"
-                    :error="editForm.errors.name"
-                    class="flex-1"
-                    v-slot="{ id }"
-                >
-                    <Input
-                        :id="id"
-                        v-model="editForm.name"
-                        data-test="group-edit-name-input"
-                        autocomplete="off"
-                    />
-                </FormField>
-                <FormField
-                    :label="$t('Group handle')"
-                    label-class="sr-only"
-                    :error="editForm.errors.slug"
-                    class="flex-1"
-                    v-slot="{ id }"
-                >
-                    <Input
-                        :id="id"
-                        v-model="editForm.slug"
-                        data-test="group-edit-slug-input"
-                        class="font-mono"
-                        autocapitalize="off"
-                        autocomplete="off"
-                        spellcheck="false"
-                    />
-                </FormField>
-                <Button
-                    type="submit"
-                    class="rounded-full"
-                    data-test="group-rename-button"
-                    :disabled="editForm.processing"
-                >
-                    {{ $t('Save') }}
-                </Button>
-            </form>
-
-            <div class="flex flex-col gap-2">
-                <span class="text-sm font-semibold">{{ $t('Members') }}</span>
-                <p
-                    v-if="editing && editing.members.length === 0"
-                    data-test="group-members-empty"
-                    class="text-sm text-muted-foreground"
-                >
-                    {{ $t('This group has no members yet.') }}
-                </p>
-                <ul v-else class="flex flex-wrap gap-2">
-                    <li
-                        v-for="member in editing?.members ?? []"
-                        :key="member.id"
-                        class="flex items-center gap-2 rounded-full border border-border py-1 pr-1 pl-2 text-sm"
-                        :data-test="`group-member-${member.id}`"
-                    >
-                        <span
-                            class="flex size-5 shrink-0 items-center justify-center rounded-full bg-primary/10 text-[8px] font-semibold text-primary"
-                            aria-hidden="true"
-                            >{{ getInitials(member.name) }}</span
-                        >
-                        <span class="truncate">{{ member.name }}</span>
-                        <Button
-                            variant="ghost"
-                            size="none"
-                            type="button"
-                            class="rounded-full p-1"
-                            :disabled="membershipForm.processing"
-                            :aria-label="
-                                $t('Remove :name from the group', {
-                                    name: member.name,
-                                })
-                            "
-                            :data-test="`group-member-remove-${member.id}`"
-                            @click="removeFromGroup(member.id)"
-                        >
-                            <X class="size-3" />
-                        </Button>
-                    </li>
-                </ul>
-            </div>
-
-            <div class="flex flex-col gap-2">
-                <Input
-                    v-model="memberSearch"
-                    data-test="group-member-search"
-                    :placeholder="$t('Add a member')"
-                    class="rounded-full"
-                />
-                <ul
-                    v-if="candidates.length > 0"
-                    class="flex flex-col gap-1"
-                    data-test="group-member-candidates"
-                >
-                    <li v-for="member in candidates" :key="member.id">
-                        <Button
-                            variant="ghost"
-                            type="button"
-                            class="w-full justify-start gap-2 rounded-lg text-sm"
-                            :disabled="membershipForm.processing"
-                            :data-test="`group-member-add-${member.id}`"
-                            @click="addToGroup(member)"
-                        >
-                            <span
-                                class="flex size-5 shrink-0 items-center justify-center rounded-full bg-primary/10 text-[8px] font-semibold text-primary"
-                                aria-hidden="true"
-                                >{{ getInitials(member.name) }}</span
-                            >
-                            {{ member.name }}
-                        </Button>
-                    </li>
-                </ul>
-            </div>
-
-            <DialogFooter>
-                <DialogClose as-child>
-                    <Button variant="outline" class="rounded-full">{{
-                        $t('Done')
-                    }}</Button>
-                </DialogClose>
-            </DialogFooter>
-        </DialogContent>
-    </Dialog>
-
-    <Dialog
-        :open="pendingRemoval !== null"
-        @update:open="(open) => !open && (pendingRemoval = null)"
-    >
-        <DialogContent data-test="group-remove-dialog">
-            <DialogHeader>
-                <DialogTitle
-                    >{{
-                        $t('Delete @:handle?', {
-                            handle: pendingRemoval?.slug ?? '',
-                        })
-                    }}
-                </DialogTitle>
-                <DialogDescription>
-                    {{
-                        $t(
-                            'The group stops being mentionable and its handle becomes available again. Messages that already mentioned it show plain text, and the notifications they sent are unaffected.',
-                        )
-                    }}
-                </DialogDescription>
-            </DialogHeader>
-            <DialogFooter>
-                <DialogClose as-child>
-                    <Button variant="outline" class="rounded-full">{{
-                        $t('Cancel')
-                    }}</Button>
-                </DialogClose>
-                <Button
-                    variant="destructive"
-                    class="rounded-full"
-                    data-test="group-remove-confirm"
-                    :disabled="removalForm.processing"
-                    @click="confirmRemoval"
-                >
-                    {{ $t('Delete group') }}
-                </Button>
-            </DialogFooter>
-        </DialogContent>
-    </Dialog>
+    <DeleteUserGroupDialog v-model:group="pendingRemoval" :team="team.slug" />
 </template>
